fix: move unit test
